requirements.txt
# =================================
# REQUIREMENTS.TXT MINIMAL POUR ML/DL
# =================================

# Core data science
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=1.0.0
scipy>=1.7.0

# Image processing
opencv-python>=4.5.0
pillow>=8.0.0

# Deep Learning
tensorflow>=2.8.0

# Boosting libraries
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0

# Hyperparameter optimization
optuna>=3.0.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.60.0

# Jupyter/notebook support
ipykernel>=6.0.0
ipython>=8.0.0
jupyter>=1.0.0

# Visualization extras
plotly>=5.0.0

# Progress and monitoring
tensorboard>=2.8.0

# Dimensionality reduction and embeddings
umap-learn>=0.5.0
torch>=1.13.0
torchvision>=0.14.0

# Additional for EDA
openpyxl>=3.0.0